ComicCrafterAI/requirements.txt
Flask==2.2.5
openai==1.12.0
replicate==0.22.0
requests==2.31.0
fpdf==1.7.2
Pillow==10.2.0
